safari-600.7-branch/LayoutTests/inspector-obsolete/http-tests/network-preflight-options-expected.txt

CONSOLE MESSAGE: XMLHttpRequest cannot load http://localhost:8000/inspector/resources/cors-target.php?id=0&deny=yes. Origin http://127.0.0.1:8000 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://localhost:8000/inspector/resources/cors-target.php?id=1&deny=yes. Origin http://127.0.0.1:8000 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://localhost:8000/inspector/resources/cors-target.php?id=3&deny=yes. Origin http://127.0.0.1:8000 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://localhost:8000/inspector/resources/cors-target.php?id=4&deny=yes. Origin http://127.0.0.1:8000 is not allowed by Access-Control-Allow-Origin.
Tests that preflight OPTIONS requests appear in Network resources

Bug 63712
0 xhr:POST http://localhost:8000/inspector/resources/cors-target.php?id=0&deny=yes
1 xhr:OPTIONS http://localhost:8000/inspector/resources/cors-target.php?id=1&deny=yes
2 xhr:OPTIONS http://localhost:8000/inspector/resources/cors-target.php?id=2
2 xhr:POST http://localhost:8000/inspector/resources/cors-target.php?id=2
3 other:POST http://localhost:8000/inspector/resources/cors-target.php?id=3&deny=yes
4 xhr:OPTIONS http://localhost:8000/inspector/resources/cors-target.php?id=4&deny=yes
5 xhr:OPTIONS http://localhost:8000/inspector/resources/cors-target.php?id=5
5 xhr:POST http://localhost:8000/inspector/resources/cors-target.php?id=5
